# openapi/nova-browser-api.json
{
  "components": {
    "schemas": {
...
      "AnalysisSubmitRequest": {
        "properties": {
          "aoi_id": {
            "format": "uuid",
            "title": "Aoi Id",
            "type": "string"
          },
          "gis_metadata": {
            "$ref": "#/components/schemas/GISMetadata"
          },
          "operation": {
            "default": "flood_mapping",
            "pattern": "^flood_mapping$",
            "title": "Operation",
            "type": "string"
          },
          "parameters": {
            "$ref": "#/components/schemas/FIRRISAnalysisParameters"
          },
          "products": {
            "items": {
              "$ref": "#/components/schemas/FIRRISProduct"
            },
            "minItems": 1,
            "title": "Products",
            "type": "array"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "aoi_id",
          "products",
          "parameters",
          "gis_metadata"
        ],
        "title": "AnalysisSubmitRequest",
        "type": "object"
      },
      "AnalysisSubmitResponse": {
        "properties": {
          "task": {
            "$ref": "#/components/schemas/TaskStatusResponse"
          }
        },
        "required": [
          "task"
        ],
        "title": "AnalysisSubmitResponse",
        "type": "object"
      },
      "ApiError": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "DurationProcessingOptions": {
        "additionalProperties": false,
        "properties": {
          "gap_policy": {
            "const": "reject",
            "default": "reject",
            "title": "Gap Policy",
            "type": "string"
          },
          "temporal_resolution_hours": {
            "exclusiveMinimum": 0.0,
            "maximum": 168.0,
            "title": "Temporal Resolution Hours",
            "type": "number"
          }
        },
        "required": [
          "temporal_resolution_hours"
        ],
        "title": "DurationProcessingOptions",
        "type": "object"
      },
...
      "EngineExecutionContract": {
        "properties": {
          "engine_key": {
            "title": "Engine Key",
            "type": "string"
          },
          "engine_name": {
            "title": "Engine Name",
            "type": "string"
          },
          "engine_version": {
            "title": "Engine Version",
            "type": "string"
          },
          "operations": {
            "items": {
              "type": "string"
            },
            "title": "Operations",
            "type": "array"
          },
          "products": {
            "items": {
              "$ref": "#/components/schemas/EngineProductDefinition"
            },
            "title": "Products",
            "type": "array"
          }
        },
        "required": [
          "engine_key",
          "engine_name",
          "engine_version",
          "operations",
          "products"
        ],
        "title": "EngineExecutionContract",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "detail": {
            "title": "Detail"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "required": [
          "detail",
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "FIRRISAnalysisParameters": {
        "additionalProperties": true,
        "description": "Typed universal workflow plus backwards-compatible direct product inputs.",
        "properties": {
          "workflow": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FIRRISSatelliteWorkflowConfig"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "FIRRISAnalysisParameters",
        "type": "object"
      },
...
      "FIRRISProduct": {
        "enum": [
          "flood_extent",
          "flood_depth",
          "flood_velocity",
          "flood_hazard",
          "flood_probability",
          "flood_duration",
          "flood_exposure",
          "flood_vulnerability",
          "flood_risk",
          "flood_susceptibility",
          "flood_hazard_zonation"
        ],
        "title": "FIRRISProduct",
        "type": "string"
      },
...
      "GISMetadata": {
        "properties": {
          "acquisition_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acquisition Date"
          },
          "bounding_box": {
            "$ref": "#/components/schemas/BoundingBox"
          },
          "crs": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Crs",
            "type": "string"
          },
          "datum": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datum"
          },
          "engine_version": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine Version"
          },
          "producer": {
            "default": "NOVA GeoRisk",
            "maxLength": 255,
            "minLength": 1,
            "title": "Producer",
            "type": "string"
          },
          "projection": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projection"
          },
          "spatial_resolution": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SpatialResolution"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "crs",
          "bounding_box"
        ],
        "title": "GISMetadata",
        "type": "object"
      },
...
      "SourceBindingValidationResponse": {
        "properties": {
          "analysis_ready": {
            "title": "Analysis Ready",
            "type": "boolean"
          },
          "executable": {
            "title": "Executable",
            "type": "boolean"
          },
          "module": {
            "enum": [
              "hazard",
              "exposure",
              "vulnerability",
              "insecurity",
              "risk",
              "resilience",
              "flood_depth",
              "flood_velocity",
              "flood_hazard_product",
              "flood_aep",
              "flood_return_period",
              "flood_duration",
              "flood_susceptibility",
              "flood_hazard_zonation"
            ],
            "title": "Module",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "source_ids": {
            "additionalProperties": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Source Ids",
            "type": "object"
          }
        },
        "required": [
          "module",
          "source_ids",
          "analysis_ready",
          "executable"
        ],
        "title": "SourceBindingValidationResponse",
        "type": "object"
      },
      "SourceBoundAnalysisRequest": {
        "additionalProperties": false,
        "properties": {
          "aoi_id": {
            "format": "uuid",
            "title": "Aoi Id",
            "type": "string"
          },
          "duration_options": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DurationProcessingOptions"
              },
              {
                "type": "null"
              }
            ]
          },
          "exposure_options": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExposureProcessingOptions"
              },
              {
                "type": "null"
              }
            ]
          },
          "hazard_options": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HazardProcessingOptions"
              },
              {
                "type": "null"
              }
            ]
          },
          "module": {
            "enum": [
              "hazard",
              "exposure",
              "vulnerability",
              "insecurity",
              "risk",
              "resilience",
              "flood_depth",
              "flood_velocity",
              "flood_hazard_product",
              "flood_aep",
              "flood_return_period",
              "flood_duration",
              "flood_susceptibility",
              "flood_hazard_zonation"
            ],
            "title": "Module",
            "type": "string"
          },
          "period": {
            "$ref": "#/components/schemas/TemporalCoverage"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "sources": {
            "additionalProperties": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Sources",
            "type": "object"
          },
          "target_grid": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RasterGrid"
              },
              {
                "type": "null"
              }
            ]
          },
          "upstream_results": {
            "additionalProperties": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Upstream Results",
            "type": "object"
          }
        },
        "required": [
          "project_id",
          "aoi_id",
          "module",
          "period"
        ],
        "title": "SourceBoundAnalysisRequest",
        "type": "object"
      },
...
      "TaskStatusResponse": {
        "properties": {
          "aoi_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aoi Id"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "engine_key": {
            "title": "Engine Key",
            "type": "string"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "error_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Summary"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "progress_pct": {
            "default": 0,
            "title": "Progress Pct",
            "type": "integer"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "result_payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Payload"
          },
          "result_reference": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResultReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatusEnum"
          },
          "task_id": {
            "format": "uuid",
            "title": "Task Id",
            "type": "string"
          },
          "task_type": {
            "title": "Task Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "task_id",
          "project_id",
          "engine_key",
          "task_type",
          "status",
          "created_at"
        ],
        "title": "TaskStatusResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "NOVA-GeoRisk Browser API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/analyses": {
      "post": {
        "operationId": "submit_analysis_api_v1_analyses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisSubmitResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authorization failed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Submit Analysis",
        "tags": [
          "Analyses"
        ]
      }
    },
    "/api/v1/analyses/engines/{engine_key}": {
      "get": {
        "operationId": "get_execution_contract_api_v1_analyses_engines__engine_key__get",
        "parameters": [
          {
            "in": "path",
            "name": "engine_key",
            "required": true,
            "schema": {
              "title": "Engine Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngineExecutionContract"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authorization failed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Get Execution Contract",
        "tags": [
          "Analyses"
        ]
      }
    },
    "/api/v1/analyses/source-bindings/contract": {
      "get": {
        "operationId": "get_source_binding_contract_api_v1_analyses_source_bindings_contract_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Source Binding Contract Api V1 Analyses Source Bindings Contract Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authorization failed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Get Source Binding Contract",
        "tags": [
          "Analyses"
        ]
      }
    },
    "/api/v1/analyses/source-bindings/validate": {
      "post": {
        "operationId": "validate_source_bindings_api_v1_analyses_source_bindings_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceBoundAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceBindingValidationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authorization failed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Validate Source Bindings",
        "tags": [
          "Analyses"
        ]
      }
    },
    "/api/v1/analyses/source-bound": {
      "post": {
        "operationId": "submit_source_bound_analysis_api_v1_analyses_source_bound_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceBoundAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisSubmitResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authorization failed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Submit Source Bound Analysis",
        "tags": [
          "Analyses"
        ]
      }
    },
...
  }
}